payments_app/api/permissions.py
from rest_framework import permissions


class IsOwnerOrAdmin(permissions.BasePermission):

    def has_object_permission(self, request, view, obj):
        if request.user and request.user.is_staff:
            return True

        is_owner = obj.order.user == request.user
        return is_owner