requirements.txt
coverage==7.6.1
Django==5.0.1
django-filter==24.3
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
phonenumbers==8.13.45
shippo==3.7.0
stripe==10.12.0